RoundTwo_MoveToSqlServer
Altered the sql statements to work with Sql Server instead of Sybase
Advantage. After creating a service that replicates the commit data to
sql server once a minute was able to change the application to utilize
sql server instead. This helps with the short number of licenses
available for Advantage. It also increases running performance by about
3 fold.